♻️ refactor: migrate from js/wasmJs targets to unified web target
- Replace jsMain and wasmJsMain source sets with webMain
- Consolidate web-specific code into webMain source set
- Update build configurations to use new web target
- Rename commonWebMain files to webMain for consistency
- Add webpack configuration for watch mode
- Update dependencies and build scripts for unified web target
