+++ build-src/plugins/multiplatform/src/main/kotlin/my/kmp-library.gradle.kts
@@ -83,12 +83,6 @@ kotlin {
     }
     jvmMain.get().dependsOn(commonJvmMain)
     androidMain.get().dependsOn(commonJvmMain)
-
-    val commonWebMain by creating {
-      dependsOn(commonMain.get())
-    }
-    jsMain.get().dependsOn(commonWebMain)
-    wasmJsMain.get().dependsOn(commonWebMain)
   }
 }
