+++ conduit-frontend/app-android/build.gradle.kts
@@ -11,11 +11,6 @@ kotlin {
       implementation(project(":frontend-decompose-logic"))
       implementation(project(":frontend-compose-ui"))
 
-//      implementation(compose.runtime)
-//      implementation(compose.ui)
-//      implementation(compose.foundation)
-//      implementation(compose.material3)
-//      implementation(compose.components.resources)
       implementation(compose.components.uiToolingPreview)
       implementation(compose.uiTooling)
       implementation(compose.preview)
