+++ conduit-frontend/app-web/build.gradle.kts
@@ -1,31 +1,17 @@
 import org.jetbrains.kotlin.gradle.ExperimentalWasmDsl
-import org.jetbrains.kotlin.gradle.targets.js.webpack.KotlinWebpackConfig
 
 plugins {
   id("my.cmp-app")
 }
 
 kotlin {
-  js(IR) {
-    browser()
+  js {
+    browser {}
     binaries.executable()
   }
   @OptIn(ExperimentalWasmDsl::class)
   wasmJs {
-    browser {
-      // copied from KMP wizard
-      val rootDirPath = project.rootDir.path
-      val projectDirPath = project.projectDir.path
-      commonWebpackConfig {
-        devServer = (devServer ?: KotlinWebpackConfig.DevServer()).apply {
-          static = (static ?: mutableListOf()).apply {
-            // Serve sources to debug inside browser
-            add(rootDirPath)
-            add(projectDirPath)
-          }
-        }
-      }
-    }
+    browser {}
     binaries.executable()
   }
 
@@ -35,12 +21,8 @@ kotlin {
       implementation(project(":frontend-compose-ui"))
 
       implementation(project.dependencies.platform(libs.dev.frontend.kotlinWrapper))
-    }
-    jsMain.dependencies {
+      implementation(libs.dev.frontend.kotlinxBrowser)
       implementation("org.jetbrains.kotlin-wrappers:kotlin-browser")
-    }
-    wasmJsMain.dependencies {
-      implementation(libs.dev.frontend.browserWasm)
     }
   }
 }
