--- conduit-frontend/app-web/src/wasmJsMain/kotlin/mikufan/cx/conduit/frontend/app/wasm/setup.kt
@@ -1,11 +1,0 @@
-package mikufan.cx.conduit.frontend.app.wasm
-
-import kotlinx.browser.document
-import kotlinx.browser.window
-
-fun setupBrowserUI() {
-  document.body!!.apply {
-    setAttribute("width", "${window.innerWidth}")
-    setAttribute("height", "${window.innerHeight}")
-  }
-}
