--- conduit-frontend/app-web/src/wasmJsMain/resources/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <script src="skiko.js"></script>
-    <title>Loading Conduit ...</title> <!-- will be changed by our code -->
-    <!--
-          This is important for mobile devices. See here for more information:
-          https://developer.mozilla.org/en-US/docs/Web/CSS/Viewport_concepts#mobile_viewports
-        -->
-    <meta name="viewport" content="width=device-width">
-    <link type="text/css" rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <canvas id="ConduitCanvas"></canvas>
-    <script type="application/javascript" src="app-web.js"></script> <!-- the name here corresponds to the gradle module name or wasmJs {} setup -->
-</body>
-</html>
