+++ conduit-frontend/app-web/src/webMain/resources/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+    <title>Loading Conduit ...</title> <!-- will be changed by our code -->
+    <link type="text/css" rel="stylesheet" href="styles.css">
+    <script type="application/javascript" src="app-web.js"></script> <!-- the name here corresponds to the gradle module name or Js/wasmJs {} setup -->
+</head>
+<body> <!-- In the new ComposeViewport with a11y support, let's just use the empty body -->
+</body>
+</html>
