+++ conduit-frontend/app-web/src/webMain/resources/styles.css
@@ -5,7 +5,3 @@ html, body {
     padding: 0;
     overflow: hidden;
 }
-
-#ConduitCanvas {
-    outline: none;
-}
